Add album field to Music model
A new field 'album' was added to the Music model to store the relationship between a music track and its album. This will enrich the information available in each music object and enhance the user's search experience. This required creating a new migration. Also, unnecessary whitespaces were removed for clean coding.