Remove artist reference from Music model, fetch from Album instead
Changes are done to remove the redundancy in data storage where the artist's information was stored both in music and album models. The artist's data in music model is removed, and instead, it's fetched from album model which is associated with the music. This change simplifies the data structure and makes it more efficient by eliminating redundant fields. All the instances in views.py and home.html that were referencing musicItem.artist now have been updated to fetch the artist details from musicItem.album.artist.