Add model tests for Artists, Albums, and Music
Added a ModelsTest class to provide thorough model testing for the Artist, Album, and Music models. The tests include creation of instances and checking their inclusion in the database, as well as testing relationships between models (e.g., album belongs to artist, music belongs to album). This ensures the correctness and integrity of the model data.